Parse dates and move http.py -> client.py
