+++ tests/test_xero_future_dates_no_data.py
@@ -4,7 +4,7 @@ from singer import utils
 import tap_tester.menagerie as menagerie
 import tap_tester.runner as runner
 
-from .base import XeroScenarioBase
+from base import XeroScenarioBase
 
 class XeroFutureDatesNoData(XeroScenarioBase):
     def name(self):
